+++ cmd/template/cluster/flag.go
@@ -15,55 +15,74 @@ const (
 	flagProvider = "provider"
 
 	// AWS only.
-	flagExternalSNAT = "external-snat"
-	flagPodsCIDR     = "pods-cidr"
+	flagExternalSNAT       = "external-snat"
+	flagPodsCIDR           = "pods-cidr"
+	flagControlPlaneSubnet = "control-plane-subnet"
 
 	// Common.
-	flagClusterID = "cluster-id"
-	flagMasterAZ  = "master-az"
-	flagName      = "name"
-	flagOutput    = "output"
-	flagOwner     = "owner"
-	flagRelease   = "release"
-	flagLabel     = "label"
+	flagClusterID      = "cluster-id"
+	flagControlPlaneAZ = "control-plane-az"
+	flagMasterAZ       = "master-az" // TODO: Remove some time after August 2021
+	flagName           = "name"
+	flagOutput         = "output"
+	flagOwner          = "owner"
+	flagRelease        = "release"
+	flagLabel          = "label"
 )
 
 type flag struct {
 	Provider string
 
 	// AWS only.
-	ExternalSNAT bool
-	PodsCIDR     string
+	ControlPlaneSubnet string
+	ExternalSNAT       bool
+	PodsCIDR           string
 
 	// Common.
-	ClusterID string
-	MasterAZ  []string
-	Name      string
-	Output    string
-	Owner     string
-	Release   string
-	Label     []string
+	ClusterID      string
+	ControlPlaneAZ []string
+	MasterAZ       []string
+	Name           string
+	Output         string
+	Owner          string
+	Release        string
+	Label          []string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Provider, flagProvider, "", "Installation infrastructure provider.")
 
 	// AWS only.
+	cmd.Flags().StringVar(&f.ControlPlaneSubnet, flagControlPlaneSubnet, "", "Subnet used for the Control Plane.")
 	cmd.Flags().BoolVar(&f.ExternalSNAT, flagExternalSNAT, false, "AWS CNI configuration.")
 	cmd.Flags().StringVar(&f.PodsCIDR, flagPodsCIDR, "", "CIDR used for the pods.")
 
 	// Common.
 	cmd.Flags().StringVar(&f.ClusterID, flagClusterID, "", "User-defined cluster ID.")
-	cmd.Flags().StringSliceVar(&f.MasterAZ, flagMasterAZ, []string{}, "Workload master node availability zone.")
+	cmd.Flags().StringSliceVar(&f.ControlPlaneAZ, flagControlPlaneAZ, nil, "Availability zone(s) to use by control plane nodes.")
+	cmd.Flags().StringSliceVar(&f.MasterAZ, flagMasterAZ, nil, "Replaced by --control-plane-az.")
 	cmd.Flags().StringVar(&f.Name, flagName, "", "Workload cluster name.")
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", "File path for storing CRs.")
 	cmd.Flags().StringVar(&f.Owner, flagOwner, "", "Workload cluster owner organization.")
 	cmd.Flags().StringVar(&f.Release, flagRelease, "", "Workload cluster release. If not given, this remains empty for defaulting to the most recent one via the Management API.")
 	cmd.Flags().StringSliceVar(&f.Label, flagLabel, nil, "Workload cluster label.")
+
+	// TODO: Remove the flag completely some time after August 2021
+	_ = cmd.Flags().MarkDeprecated(flagMasterAZ, "please use --control-plane-az.")
 }
 
 func (f *flag) Validate() error {
 	var err error
+
+	// TODO: Remove the flag completely some time after August 2021
+	if len(f.MasterAZ) > 0 {
+		if len(f.ControlPlaneAZ) > 0 {
+			return microerror.Maskf(invalidFlagError, "--control-plane-az and --master-az cannot be combined")
+		}
+
+		f.ControlPlaneAZ = f.MasterAZ
+		f.MasterAZ = nil
+	}
 
 	if f.Provider != key.ProviderAWS && f.Provider != key.ProviderAzure {
 		return microerror.Maskf(invalidFlagError, "--%s must be either aws or azure", flagProvider)
@@ -91,6 +110,13 @@ func (f *flag) Validate() error {
 			return microerror.Maskf(invalidFlagError, "--%s must only contain alphanumeric characters, and start with a letter", flagClusterID)
 		}
 
+		if f.ControlPlaneSubnet != "" {
+			matchedSubnet, err := regexp.MatchString("^20|21|22|23|24|25$", f.ControlPlaneSubnet)
+			if err == nil && !matchedSubnet {
+				return microerror.Maskf(invalidFlagError, "--%s must be a valid subnet size (20, 21, 22, 23, 24 or 25)", flagControlPlaneSubnet)
+			}
+		}
+
 		return nil
 	}
 	// Validate name for non-aws clusters.
@@ -110,12 +136,12 @@ func (f *flag) Validate() error {
 		// Validate Master AZs.
 		switch f.Provider {
 		case key.ProviderAWS:
-			if len(f.MasterAZ) != 0 && len(f.MasterAZ) != 1 && len(f.MasterAZ) != 3 {
-				return microerror.Maskf(invalidFlagError, "--%s must be set to either one or three availability zone names", flagMasterAZ)
+			if len(f.ControlPlaneAZ) != 0 && len(f.ControlPlaneAZ) != 1 && len(f.ControlPlaneAZ) != 3 {
+				return microerror.Maskf(invalidFlagError, "--%s must be set to either one or three availability zone names", flagControlPlaneAZ)
 			}
 		case key.ProviderAzure:
-			if len(f.MasterAZ) > 1 {
-				return microerror.Maskf(invalidFlagError, "--%s supports one availability zone only", flagMasterAZ)
+			if len(f.ControlPlaneAZ) > 1 {
+				return microerror.Maskf(invalidFlagError, "--%s supports one availability zone only", flagControlPlaneAZ)
 			}
 		}
 	}