+++ cmd/template/cluster/provider/aws.go
@@ -4,6 +4,7 @@ import (
 	"io"
 	"text/template"
 
+	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
 	"github.com/giantswarm/microerror"
 	"sigs.k8s.io/yaml"
@@ -15,9 +16,10 @@ func WriteAWSTemplate(out io.Writer, config ClusterCRsConfig) error {
 	var err error
 
 	crsConfig := v1alpha2.ClusterCRsConfig{
-		ClusterID:      config.ClusterID,
+		ClusterID: config.ClusterID,
+
 		ExternalSNAT:   config.ExternalSNAT,
-		MasterAZ:       config.MasterAZ,
+		MasterAZ:       config.ControlPlaneAZ,
 		Description:    config.Description,
 		PodsCIDR:       config.PodsCIDR,
 		Owner:          config.Owner,
@@ -28,6 +30,10 @@ func WriteAWSTemplate(out io.Writer, config ClusterCRsConfig) error {
 	crs, err := v1alpha2.NewClusterCRs(crsConfig)
 	if err != nil {
 		return microerror.Mask(err)
+	}
+
+	if config.ControlPlaneSubnet != "" {
+		crs.AWSCluster.Annotations[annotation.AWSSubnetSize] = config.ControlPlaneSubnet
 	}
 
 	clusterCRYaml, err := yaml.Marshal(crs.Cluster)