+++ cmd/template/nodepool/runner.go
@@ -53,6 +53,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 			ClusterID:                           r.flag.ClusterID,
 			Description:                         r.flag.NodepoolName,
 			VMSize:                              r.flag.AzureVMSize,
+			MachineDeploymentSubnet:             r.flag.MachineDeploymentSubnet,
 			NodesMax:                            r.flag.NodesMax,
 			NodesMin:                            r.flag.NodesMin,
 			OnDemandBaseCapacity:                r.flag.OnDemandBaseCapacity,