+++ CHANGELOG.md
@@ -7,13 +7,35 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Replace AppCatalog CRD with new namespace scoped Catalog CRD.
+
+## [1.30.0] - 2021-06-29
+
+### Changed
+
+- Extend `template app` to only output required fields, the flag `--defaulting-enabled`
+can be set to false to disable this.
+
+## [1.29.2] - 2021-06-17
+
+- In the `template cluster` command, the flag `--control-plane-az` is replacing `--master-az`.
+
+## [1.29.1] - 2021-06-16
+
+- Modify the AWS subnet validation for machine deployments.
+
+## [1.29.0] - 2021-06-15
+
 ### Added
 
+- Add the AWS subnet annotations into the `template` command.
 - Limit the time allowed for the `login` command to call the authentication proxy to one minute.
 
 ### Changed
 
-- Replace AppCatalog CRD with new namespace scoped Catalog CRD.
+- Updated terminology to use "control plane nodes" instead of "master nodes".
 
 ## [1.28.0] - 2021-05-11
 
@@ -347,7 +369,11 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.28.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.30.0...HEAD
+[1.30.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.2...v1.30.0
+[1.29.2]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.1...v1.29.2
+[1.29.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.0...v1.29.1
+[1.29.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.28.0...v1.29.0
 [1.28.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.27.1...v1.28.0
 [1.27.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.27.0...v1.27.1
 [1.27.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.26.0...v1.27.0